requirements.txt
beautifulsoup4==4.8.2
celery==4.4.2
Django==3.0.4
djangorestframework==3.11.0
django_celery_beat==2.0.0
lxml==4.5.0
numpy==1.18.2
psycopg2==2.8.4
redis==3.2.0
requests==2.23.0
scikit-learn==0.22.2
